Implementing UI Class
Implemented UserInterface class to handle all console prints the user should see.
